Bugfix
Made Sure to Eliminate Any Commands That Changed Directory. This Ensures That no Matter Where The Install Script is, it Can Properly Move All Dotfiles to The Right Locations Without Error of Not Finding Them in The Current Working Directory. Can't Believe I Was Stupid Enough to Miss Such A Major Bug :P